+++ Cargo.toml
@@ -6,5 +6,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+eframe = "0.24.1"
 probability = "0.18.0"
 rayon = "1.5"
